Added filling ConfigurationsDialog table with data
Model for table is created and used, but there is no way to parser
arbitrary data file for now. Such functionality is out of scope of
GUI/visualization task, and it would probably be useful in more
situations. Once this is in place, this piece of visualization task can
be completed.